Splitting project further
