+++ packages/models-library/src/models_library/projects_ui.py
@@ -2,16 +2,17 @@ from typing import Dict, Optional
 
 from pydantic import BaseModel, Extra, Field, constr
 from .basic_regex import UUID_RE
-
+from .project_nodes import Node, Position
 NodeID = constr(regex=UUID_RE)
 
+# Pydantic does not support exporting a jsonschema with Dict keys being something else than a str
+# this is a regex for having uuids of type: 8-4-4-4-12 digits
+_NodeIDForDict = constr(
+    regex=r"^[0-9a-fA-F]{8}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{4}-?[0-9a-fA-F]{12}$"
+)
+Workbench = Dict[_NodeIDForDict, Node]
 
-class Position(BaseModel):
-    x: int = Field(..., description="The x position", example=["12"])
-    y: int = Field(..., description="The y position", example=["15"])
 
-    class Config:
-        extra = Extra.forbid
 
 
 class WorkbenchUI(BaseModel):
@@ -28,6 +29,11 @@ class Slideshow(BaseModel):
         extra = Extra.forbid
 
 
+
 class StudyUI(BaseModel):
-    workbench: Optional[Dict[NodeID, WorkbenchUI]] = Field(None)
-    slideshow: Optional[Dict[NodeID, Slideshow]] = Field(None)
+    workbench: Optional[Dict[_NodeIDForDict, WorkbenchUI]] = Field(None)
+    slideshow: Optional[Dict[_NodeIDForDict, Slideshow]] = Field(None)
+    current_node_id: Optional[NodeID] = Field(alias="currentNodeId")
+
+    class Config:
+        extra = Extra.allow
